crewProject/requirements.txt
crewai
crewai-tools
langchain-google-genai
python-dotenv
google-generativeai
beautifulsoup4
requests
googlesearch-python

